create lib/sync
